requirements.txt
asgiref==3.6.0
channels==4.0.0
Django==4.1.4
django-admin-volt==1.0.4
django-mathfilters==1.0.0
django-ratelimit==4.0.0
django-shopping-cart==0.1
django-tinymce==3.5.0
Pillow==9.3.0
six==1.16.0
sqlparse==0.4.3
